Rework Movement
Instead of using the old system that had all of the ways of movement
using the same function for applying the force, we apply the force
differently depending on the mode of movement. This along with some
smaller fixes makes sliding movement and aerial movement much smoother
to use and implement.

Also adds bhopping